fix(tts): switch KV cache dtype fp16 → fp32 for q4 LM
The q4 LM expects fp32 KV cache tensors (the f16 suffix in q4f16 marked
activation precision — q4 has fp32 activations, including the past/present
key-value pairs). Workers previously hardcoded float16 KV cache because
the q4f16 export required it. Mismatch produced:

  Unexpected input data type. Actual: (tensor(float16)), expected: (tensor(float))

at the first session.run on warmup, then again on every synth attempt.

Verified upstream q4 LM ONNX inputs via onnx.load:
  inputs_embeds:           dtype=float32
  attention_mask:          dtype=int64
  past_key_values.X.{k,v}: dtype=float32
  logits:                  dtype=float32
  present.X.{k,v}:         dtype=float32

Updated both worker init paths (initial empty KV, warmup, CFG branches)
to allocate fp32 KV instead. Float32Array(0) replaces Uint16Array(0) at
each construction site.

Memory note: fp32 KV cache is 2x the fp16 size — at 50 tokens × 30 layers
with CFG batch=2, ~12 MB vs ~6 MB. Still fits in M5 GPU memory budget by
a wide margin.

+++ public/tts-gpu-worker.js
@@ -704,8 +704,8 @@ async function warmupLanguageModel() {
     attention_mask: new ort.Tensor("int64", BigInt64Array.from({ length: batch * totalLen }, () => 1n), [batch, totalLen]),
   };
   for (let i = 0; i < NUM_LAYERS; i++) {
-    passA[`past_key_values.${i}.key`] = new ort.Tensor("float16", new Uint16Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
-    passA[`past_key_values.${i}.value`] = new ort.Tensor("float16", new Uint16Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
+    passA[`past_key_values.${i}.key`] = new ort.Tensor("float32", new Float32Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
+    passA[`past_key_values.${i}.value`] = new ort.Tensor("float32", new Float32Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
   }
   const resultA = await languageModelSession.run(passA);
 
@@ -833,8 +833,8 @@ async function handleSynthesize(text, speakerData, id, languageId, exaggeration 
     attention_mask: new ort.Tensor("int64", BigInt64Array.from({ length: batch * totalLen }, () => 1n), [batch, totalLen]),
   };
   for (let i = 0; i < NUM_LAYERS; i++) {
-    lmInputs[`past_key_values.${i}.key`] = new ort.Tensor("float16", new Uint16Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
-    lmInputs[`past_key_values.${i}.value`] = new ort.Tensor("float16", new Uint16Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
+    lmInputs[`past_key_values.${i}.key`] = new ort.Tensor("float32", new Float32Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
+    lmInputs[`past_key_values.${i}.value`] = new ort.Tensor("float32", new Float32Array(0), [batch, NUM_HEADS, 0, HEAD_DIM]);
   }
 
   const generatedTokens = [START_SPEECH_TOKEN];
